+++ src/modules/Elsa.MassTransit.AzureServiceBus/Commands/CleanupSubscriptions.cs
@@ -1,8 +1,8 @@
 using Elsa.Mediator.Contracts;
 
-namespace Elsa.MassTransit.AzureServiceBus.Notifications;
+namespace Elsa.MassTransit.AzureServiceBus.Commands;
 
 /// <summary>
 /// Notification to clean up the Azure Service Bus subscriptions.
 /// </summary>
-public record CleanupSubscriptions : INotification;
+public record CleanupSubscriptions : ICommand;
